+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next version
+## 2.4.0 - December 22, 2024
 
 ### 💥 Breaking changes